Lagt til 9 tester(veldig snart ferdig med 10. unit test)
og 1 integrasjonstest

Lagt til testtag i funksjonen Sikt_FinnTurer_card, for å teste bilde

+++ app/src/main/java/com/example/in2000_prosjekt/ui/data/ImplementedWeatherRepository.kt
@@ -118,7 +118,7 @@ class ImplementedWeatherRepository : WeatherRepository {
     override suspend fun getFrost(latitude: String, longitude: String): FrostInfo {
 
         val frost = dataFrost.fetchFrostTemp(elements, referencetime, source)  //hardkoded parameterne, fiks dette
-        val frostPolygon = dataFrost.fetchApiSvarkoordinater(latitude, longitude)
+        val frostPolygon = dataFrost.fetchApiSvarkoordinater(latitude.toDouble(), longitude.toDouble())
 
         val typeFrost = frost.type
         val long = frostPolygon.data?.get(0)?.geometry?.coordinates?.get(0)
